Dedicated pod annotations for /dev/nvidia.. devices
Pods which have a gpu cores map specified in Triad config receive
these mappings in a form of annotation list. Annotations can be used
to expose individual nvidia devices as hostPath volume objects to the
k8s pod (instead of just exposing whole /dev contents from the host)

The example form of the annotation is:

sigproc.viasat.io/nhd_gpu_devices.nvidia0: 3

... where the device name as seen on the pod is the last segment of the label
here - "nvidia0" and the physical gpu core name is the value of the annotation -
in this case "3".