Add support for extra NIC label and marking NIC as "backup"
This patch adds support for parsing extra NIC label which is
supposed to be adde as last segment of the label name by NFD.

If this last segment of the label is set to "bkp" - then the
NIC is labeled as a spare/failover NIC to be allocated whenever
dual_port option is set. NIC-s with this label are excluded
from general NHD matchmaking process.